# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.IngredientResponse:
    properties:
      created_at:
        type: string
      current_stock:
        type: number
      id:
        type: integer
      is_low:
        type: boolean
      min_threshold:
        type: number
      name:
        type: string
      unit:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.LinkIngredientRequest:
    properties:
      ingredient_id:
        type: integer
      quantity:
        type: number
    required:
    - ingredient_id
    - quantity
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuItemIngredientResponse:
    properties:
      ingredient:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.IngredientResponse'
      ingredient_id:
        type: integer
      quantity:
        type: number
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.UpdateLinkItemRequest:
    properties:
      quantity:
        type: number
    required:
    - quantity
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: https://geritcht-restaurant.onrender.com
info:
  contact:
    email: alabifathiu@yahoo.com
    name: Alabi Fathiu
    url: http://linkedin.com/in/fathiu-alabi
  description: A modern Restaurant API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Geritcht Restaurant API
  version: "1.0"
paths:
...
  /recipes/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all ingredients linked to a menu item's recipe.
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recipe retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuItemIngredientResponse'
                  type: array
              type: object
        "400":
          description: Invalid menu ID
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Menu not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get recipe
      tags:
      - Recipes
    post:
      consumes:
      - application/json
      description: Link an ingredient to a menu item's recipe with the specified quantity.
        Admin access required.
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recipe ingredient details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.LinkIngredientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Ingredient added to recipe successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuItemIngredientResponse'
              type: object
        "400":
          description: Invalid request data, ingredient ID, or quantity
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Ingredient not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "409":
          description: Ingredient already linked to this recipe
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Add ingredient to a recipe
      tags:
      - Recipes
  /recipes/{id}/{ingredientID}:
    delete:
      consumes:
      - application/json
      description: Remove an ingredient from a menu item's recipe. Admin access required.
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ingredient ID
        in: path
        name: ingredientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recipe deleted successfully
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Recipe or ingredient not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Remove ingredient from recipe
      tags:
      - Recipes
    patch:
      consumes:
      - application/json
      description: Update the quantity or details of an ingredient linked to a menu
        item's recipe. Admin access required.
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ingredient ID
        in: path
        name: ingredientID
        required: true
        type: integer
      - description: Updated recipe ingredient details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.UpdateLinkItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Recipe updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuItemIngredientResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Recipe or ingredient not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Update recipe ingredient
      tags:
      - Recipes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"